requirements.txt
# Data Fetching & API
requests>=2.32.0
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Database & Logic
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
pandas>=2.2.0

# AI & LLM Orchestration
# We remove the exact versions here to let pip resolve the 'jiter' conflict
openai>=2.0.0
instructor>=1.7.0
groq>=0.11.0